move store into folder reduxStore
