+++ src/reduxStore/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from "@reduxjs/toolkit";
-import { settingsReducer } from "./Components/Controls/ControlSlice";
-import { gameReducer } from "./Components/Game/GameSlice";
+import { settingsReducer } from "../Components/Controls/ControlSlice";
+import { gameReducer } from "../Components/Game/GameSlice";
 
 export const store = configureStore({
   reducer: {
